remove dupes reduce hdri size
