+++ src/App/App.ts
@@ -1,11 +1,10 @@
-import { Clock, PCFSoftShadowMap, ReinhardToneMapping, Vector2, WebGLRenderer } from 'three';
+import { Clock, ReinhardToneMapping, Vector2, WebGLRenderer } from 'three';
 import { EffectComposer } from 'three/examples/jsm/postprocessing/EffectComposer';
 import { RenderPass } from 'three/examples/jsm/postprocessing/RenderPass';
 import { UnrealBloomPass } from 'three/examples/jsm/postprocessing/UnrealBloomPass';
 
 import { OutputPass } from 'three/examples/jsm/postprocessing/OutputPass';
 import { INTERSECTION_THRESHOLD } from './constants';
-import { AppScene } from './Scene';
 import { isMobile } from './utils';
 
 const IS_DESKTOP = !isMobile();
@@ -20,17 +19,17 @@ export class App {
   private controls;
   private intersectionObserver;
 
-  constructor(canvas) {
+  constructor(canvas, scene, initProps = {}) {
     this.clock = new Clock();
-    this.init(canvas);
+    this.init(canvas, scene, initProps);
     this.onResize();
     this.onShow();
   }
 
   public get dimensions() {
-    // const container = this.renderer.domElement.parentElement;
+    const container = this.renderer.domElement.parentElement;
 
-    const container = document.querySelector('.planet__img');
+    // const container = document.querySelector('.planet__img');
 
     return [container.clientWidth, container.clientHeight];
   }
@@ -40,14 +39,12 @@ export class App {
     return w / h;
   }
 
-  private init(canvas) {
+  private init(canvas, sceneClass, props: any = {}) {
     this.renderer = new WebGLRenderer({
       alpha: true,
       antialias: true,
       canvas,
     });
-    this.renderer.shadowMap.enabled = true;
-    this.renderer.shadowMap.type = PCFSoftShadowMap;
 
     const [w, h] = this.dimensions;
 
@@ -56,7 +53,7 @@ export class App {
 
     this.composer = new EffectComposer(this.renderer);
 
-    this.scene = new AppScene(this.renderer, this);
+    this.scene = new sceneClass(this.renderer, this);
 
     const { scene, camera } = this.scene;
 
@@ -67,9 +64,13 @@ export class App {
       0.85,
     );
 
-    bloomPass.threshold = 0.2;
-    bloomPass.strength = 0.05;
-    bloomPass.radius = 0.05;
+    // bloomPass.threshold = 0.2;
+    // bloomPass.strength = 0.05;
+    // bloomPass.radius = 0.05;
+
+    bloomPass.threshold = props.bloomPassThreshold;
+    bloomPass.strength = props.bloomPassStrength;
+    bloomPass.radius = props.bloomPassRadius;
 
     const outputPass = new OutputPass(ReinhardToneMapping, 1.1);
