--- src/App/shaders/overlayGradient/index.ts
@@ -1,16 +1,0 @@
-// import base from './shader.vert';
-// import gradient from './shader.frag';
-// import { GRADIENT_COLORS, GRADIENT_STEP, SATURATION } from 'App/constants';
-//
-// export default {
-//   uniforms: {
-//     tDiffuse:   { value: null },
-//     // color1:     { value: GRADIENT_COLORS[0] },
-//     // color2:     { value: GRADIENT_COLORS[1] },
-//     step:       { value: GRADIENT_STEP },
-//     saturation: { value: SATURATION },
-//   },
-//   vertexShader: base,
-//   fragmentShader: gradient,
-//
-// };
