--- src/App/shaders/overlayGradient/shader.vert
@@ -1,5 +1,0 @@
-varying vec2 vUv;
-void main() {
-  vUv = uv;
-  gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
-}
